+++ taskfile/platforms.go
@@ -2,10 +2,11 @@ package taskfile
 
 import (
 	"fmt"
-	"runtime"
 	"strings"
 
 	"gopkg.in/yaml.v3"
+
+	"github.com/go-task/task/v3/internal/goext"
 )
 
 // Platform represents GOOS and GOARCH values
@@ -14,72 +15,47 @@ type Platform struct {
 	Arch string
 }
 
-// ParsePlatform takes a string representing an OS/Arch combination (or either on their own)
-// and parses it into the Platform struct. It returns an error if the input string is invalid.
-// Valid combinations for input: OS, Arch, OS/Arch
-func (p *Platform) ParsePlatform(input string) error {
-	// tidy up input
-	platformString := strings.ToLower(strings.TrimSpace(input))
-	splitValues := strings.Split(platformString, "/")
-	if len(splitValues) > 2 {
-		return fmt.Errorf("task: Invalid OS/Arch provided: %s", input)
-	}
-	err := p.parseOsOrArch(splitValues[0])
-	if err != nil {
-		return err
-	}
-	if len(splitValues) == 2 {
-		return p.parseArch(splitValues[1])
-	}
-	return nil
+type ErrInvalidPlatform struct {
+	Platform string
 }
 
-// supportedOSes is a list of supported OSes
-var supportedOSes = map[string]struct{}{
-	"windows": {},
-	"darwin":  {},
-	"linux":   {},
-	"freebsd": {},
-}
-
-func isSupportedOS(input string) bool {
-	_, exists := supportedOSes[input]
-	return exists
-}
-
-// supportedArchs is a list of supported architectures
-var supportedArchs = map[string]struct{}{
-	"amd64": {},
-	"arm64": {},
-	"386":   {},
-}
-
-func isSupportedArch(input string) bool {
-	_, exists := supportedArchs[input]
-	return exists
-}
-
-// MatchesCurrentPlatform returns true if the platform matches the current platform
-func (p *Platform) MatchesCurrentPlatform() bool {
-	return (p.OS == "" || p.OS == runtime.GOOS) &&
-		(p.Arch == "" || p.Arch == runtime.GOARCH)
+func (err *ErrInvalidPlatform) Error() string {
+	return fmt.Sprintf(`task: Invalid platform "%s"`, err.Platform)
 }
 
 // UnmarshalYAML implements yaml.Unmarshaler interface.
 func (p *Platform) UnmarshalYAML(node *yaml.Node) error {
 	switch node.Kind {
-
 	case yaml.ScalarNode:
 		var platform string
 		if err := node.Decode(&platform); err != nil {
 			return err
 		}
-		if err := p.ParsePlatform(platform); err != nil {
+		if err := p.parsePlatform(platform); err != nil {
 			return err
 		}
 		return nil
 	}
 	return fmt.Errorf("yaml: line %d: cannot unmarshal %s into platform", node.Line, node.ShortTag())
+}
+
+// parsePlatform takes a string representing an OS/Arch combination (or either on their own)
+// and parses it into the Platform struct. It returns an error if the input string is invalid.
+// Valid combinations for input: OS, Arch, OS/Arch
+func (p *Platform) parsePlatform(input string) error {
+	splitValues := strings.Split(input, "/")
+	if len(splitValues) > 2 {
+		return &ErrInvalidPlatform{Platform: input}
+	}
+	if err := p.parseOsOrArch(splitValues[0]); err != nil {
+		return &ErrInvalidPlatform{Platform: input}
+	}
+	if len(splitValues) == 2 {
+		if err := p.parseArch(splitValues[1]); err != nil {
+			return &ErrInvalidPlatform{Platform: input}
+		}
+	}
+	return nil
 }
 
 // parseOsOrArch will check if the given input is a valid OS or Arch value.
@@ -88,16 +64,17 @@ func (p *Platform) parseOsOrArch(osOrArch string) error {
 	if osOrArch == "" {
 		return fmt.Errorf("task: Blank OS/Arch value provided")
 	}
-	if isSupportedOS(osOrArch) {
+	if goext.IsKnownOS(osOrArch) {
 		p.OS = osOrArch
 		return nil
 	}
-	if isSupportedArch(osOrArch) {
+	if goext.IsKnownArch(osOrArch) {
 		p.Arch = osOrArch
 		return nil
 	}
 	return fmt.Errorf("task: Invalid OS/Arch value provided (%s)", osOrArch)
 }
+
 func (p *Platform) parseArch(arch string) error {
 	if arch == "" {
 		return fmt.Errorf("task: Blank Arch value provided")
@@ -105,7 +82,7 @@ func (p *Platform) parseArch(arch string) error {
 	if p.Arch != "" {
 		return fmt.Errorf("task: Multiple Arch values provided")
 	}
-	if isSupportedArch(arch) {
+	if goext.IsKnownArch(arch) {
 		p.Arch = arch
 		return nil
 	}